psalm 1

+++ src/Models/Log.php
@@ -35,7 +35,7 @@ class Log extends Model
      */
     public function prunable(): \Illuminate\Database\Eloquent\Builder
     {
-        $days = config('footsteps.prune_logs_after_days', 100);
+        $days = (int)config('footsteps.prune_logs_after_days', 100);
 
         return static::where('created_at', '<=', now()->subDays($days));
     }
